migrate the methods
